backend/requirements.txt
uWSGI==2.0.20
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.5.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0